requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.20.0
wandb>=0.15.0
Pillow>=9.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
tqdm>=4.64.0
sentencepiece>=0.1.99
protobuf>=3.20.0 